# frinq-mobile/docs/openapi-snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReauthTokenResponse": {
        "properties": {
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "reauth_token": {
            "title": "Reauth Token",
            "type": "string"
          }
        },
        "required": [
          "reauth_token",
          "expires_in"
        ],
        "title": "ReauthTokenResponse",
        "type": "object"
      },
...
      "RegisterRequest": {
        "description": "Called once on first login after Supabase Auth issues a JWT.\n\nThe Supabase UID and phone come from the JWT itself \u2014 only the user-supplied\nfields appear in the body.",
        "properties": {
          "age": {
            "anyOf": [
              {
                "maximum": 65.0,
                "minimum": 18.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "display_name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "gender": {
            "anyOf": [
              {
                "enum": [
                  "male",
                  "female",
                  "non_binary",
                  "other"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "max_travel_km": {
            "default": 15,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Max Travel Km",
            "type": "integer"
          },
          "ncr_zone": {
            "anyOf": [
              {
                "enum": [
                  "gurgaon",
                  "south_delhi",
                  "noida",
                  "east_delhi",
                  "west_delhi",
                  "faridabad",
                  "other_ncr"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ncr Zone"
          },
          "schedule": {
            "items": {
              "enum": [
                "weekday_morn",
                "weekday_eve",
                "saturday",
                "sunday"
              ],
              "type": "string"
            },
            "title": "Schedule",
            "type": "array"
          }
        },
        "required": [
          "display_name"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
      "ReverifyVerifyRequest": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "ReverifyVerifyRequest",
        "type": "object"
      },
...
      "UserResponse": {
        "properties": {
          "age": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "banned": {
            "default": false,
            "title": "Banned",
            "type": "boolean"
          },
          "community_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Community Slug"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "gender": {
            "anyOf": [
              {
                "enum": [
                  "male",
                  "female",
                  "non_binary",
                  "other"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "max_travel_km": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Travel Km"
          },
          "ncr_zone": {
            "anyOf": [
              {
                "enum": [
                  "gurgaon",
                  "south_delhi",
                  "noida",
                  "east_delhi",
                  "west_delhi",
                  "faridabad",
                  "other_ncr"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ncr Zone"
          },
          "onboarding_complete": {
            "default": false,
            "title": "Onboarding Complete",
            "type": "boolean"
          },
          "onboarding_state": {
            "default": "quiz_in_progress",
            "title": "Onboarding State",
            "type": "string"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "privacy_accepted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Privacy Accepted At"
          },
          "privacy_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Privacy Version"
          },
          "schedule": {
            "items": {
              "type": "string"
            },
            "title": "Schedule",
            "type": "array"
          },
          "supabase_uid": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supabase Uid"
          },
          "terms_accepted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terms Accepted At"
          },
          "terms_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terms Version"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "UserResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Frinq Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "operationId": "register_api_v1_auth_register_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/reverify/request": {
      "post": {
        "description": "Sends an OTP to the account's own stored phone only \u2014 never accepts\na phone field, so this can't be used to send an OTP to someone else's\nnumber under an authenticated session.",
        "operationId": "reverify_request_api_v1_auth_reverify_request_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Reverify Request Api V1 Auth Reverify Request Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reverify Request",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/reverify/verify": {
      "post": {
        "description": "On success, returns a single-purpose reauth token bound to this user\nID, session ID, and the account_delete action \u2014 5-minute expiry,\nsingle-use (app/core/reverify.py). Never a general-purpose login token.",
        "operationId": "reverify_verify_api_v1_auth_reverify_verify_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReverifyVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReauthTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reverify Verify",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}